... In particular, narrowing integer conversions are done only in
sampleCount::as_size_t which asserts correctness.

... And the overuse of that type is now corrected.  size_t is now used for any
number of samples that fits in memory or a block file.  sampleCount is reserved
to describe an offset into an audio file or wave track or clip, or the
difference of two such.  sampleCount is signed, but size_t of course is not.
sampleCount is 64 bits wide, but size_t is narrower (in 32 bit executables,
such as we still build).

  Hide implicit conversions sampleCount <-> floating point, -> integer
  Change sampleCount arguments, variables, return values to size_t...
  Make many conversions sampleCount->size_t are explicit and checked...
  Convert sampleCount <-> floating or -> long long explicitly ...
  Assert that sampleCount doesn't narrow converting to 3d party types
  Define sampleCount as a class, not a type alias...

+++ src/AudioIO.h
@@ -573,7 +573,7 @@ private:
    double              mSeek;
    double              mPlaybackRingBufferSecs;
    double              mCaptureRingBufferSecs;
-   long                mPlaybackSamplesToCopy;
+   size_t              mPlaybackSamplesToCopy;
    double              mMinCaptureSecsToCopy;
    bool                mPaused;
    PaStream           *mPortStreamV19;
